fixing-redux
